requirements-dev.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.7
python-dotenv==1.0.1
werkzeug==3.0.3